requirements.txt
Django>=5.0
djangorestframework
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
